refactor(agent): extract presentation tool projection
